Preserve ClickHouse histogram totals
Extend the fresh histogram table schema with optional sum and count columns, keep OTLP bucket counts as 64-bit values, and carry totals through ingestion and query paths. Widen histogram sums to doubles while preserving absent OTLP sums as null.

+++ okapi-ingester/src/main/java/org/okapi/metrics/ch/ChHistoSample.java
@@ -28,7 +28,7 @@ public class ChHistoSample {
   @SerializedName("histo_type")
   HISTO_TYPE histoType;
 
-  float sum;
+  Double sum;
   Long count;
   String unit;
